requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
pydantic==2.6.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.24.1
email-validator==2.3.0
razorpay==1.4.1
groq==0.9.0
setuptools==75.8.0